Turn LLMService into an interface IrisModel
